+++ src/Foreign/JavaScript/Types.hs
@@ -2,6 +2,7 @@
 module Foreign.JavaScript.Types where
 
 import           Control.Applicative
+import qualified Control.Exception       as E
 import           Control.Concurrent.STM  as STM
 import           Control.Concurrent.Chan as Chan
 import           Control.Concurrent.MVar
@@ -13,6 +14,7 @@ import           Data.IORef
 import           Data.Map                as Map
 import           Data.String
 import           Data.Text
+import           Data.Typeable
 import           System.IO                       (stderr)
 
 import Foreign.RemotePtr
@@ -85,6 +87,7 @@ readComm c = STM.readTQueue (commIn c)
 data ClientMsg
     = Event Coupon JSON.Value
     | Result JSON.Value
+    | Exception String
     | Quit
     deriving (Eq, Show)
 
@@ -96,6 +99,8 @@ instance FromJSON ClientMsg where
                 Event  <$> (msg .: "name") <*> (msg .: "arguments")
             "Result" ->
                 Result <$> (msg .: "contents")
+            "Exception" ->
+                Exception <$> (msg .: "contents")
             "Quit"   ->
                 return Quit
 
@@ -143,6 +148,13 @@ the function `Control.DeepSeq.force` to make sure that any exception
 is thrown before handing the message over to another thread.
 
 -}
+
+data JavaScriptException = JavaScriptException String deriving Typeable
+
+instance E.Exception JavaScriptException
+
+instance Show JavaScriptException where
+    showsPrec _ (JavaScriptException err) = showString $ "JavaScript error: " ++ err
 
 {-----------------------------------------------------------------------------
     Window & Event Loop