+++ src/Graphics/UI/Threepenny/Internal.hs
@@ -21,6 +21,7 @@ module Graphics.UI.Threepenny.Internal (
 
 import           Control.Applicative                   (Applicative)
 import           Control.Monad
+import           Control.Monad.Catch
 import           Control.Monad.Fix
 import           Control.Monad.IO.Class
 import qualified Control.Monad.Trans.RWS.Lazy as Monad
@@ -30,7 +31,7 @@ import qualified Data.Aeson              as JSON
 import qualified Foreign.JavaScript      as JS
 import qualified Foreign.RemotePtr       as Foreign
 
-import qualified Reactive.Threepenny     as E
+import qualified Reactive.Threepenny     as RB
 
 import Foreign.JavaScript hiding (runFunction, callFunction, debug, timestamp, Window)
 
@@ -40,7 +41,7 @@ import Foreign.JavaScript hiding (runFunction, callFunction, debug, timestamp, W
 -- | The type 'Window' represents a browser window.
 data Window = Window
     { jsWindow    :: JS.Window  -- JavaScript window
-    , eDisconnect :: E.Event () -- event that happens when client disconnects
+    , eDisconnect :: RB.Event () -- event that happens when client disconnects
     , wEvents     :: Foreign.Vendor Events
                      -- events associated to 'Element's
     , wChildren   :: Foreign.Vendor ()
@@ -54,7 +55,7 @@ startGUI
     -> IO ()
 startGUI config init = JS.serve config $ \w -> do
     -- set up disconnect event
-    (eDisconnect, handleDisconnect) <- E.newEvent
+    (eDisconnect, handleDisconnect) <- RB.newEvent
     JS.onDisconnect w $ handleDisconnect ()
 
     -- make window
@@ -75,13 +76,13 @@ startGUI config init = JS.serve config $ \w -> do
 --
 -- Note: DOM Elements in a browser window that has been closed
 -- can no longer be manipulated.
-disconnect :: Window -> E.Event ()
+disconnect :: Window -> RB.Event ()
 disconnect = eDisconnect
 
 {-----------------------------------------------------------------------------
     Elements
 ------------------------------------------------------------------------------}
-type Events = String -> E.Event JSON.Value
+type Events = String -> RB.Event JSON.Value
 
 -- Reachability information for children of an 'Element'.
 -- The children of an element are always reachable from this RemotePtr.
@@ -146,7 +147,7 @@ addEvents el Window{ jsWindow = w, wEvents = wEvents } = do
             JS.runFunction w $
                 ffi "Haskell.bind(%1,%2,%3)" el name handlerPtr
 
-    events <- E.newEventsNamed initializeEvent
+    events <- RB.newEventsNamed initializeEvent
 
     -- Create new pointer and add reachability.
     Foreign.withRemotePtr el $ \coupon _ -> do
@@ -181,7 +182,7 @@ domEvent
         --   Note that the @on@-prefix is not included,
         --   the name is @click@ and so on.
     -> Element          -- ^ Element where the event is to occur.
-    -> E.Event EventData
+    -> RB.Event EventData
 domEvent name el = elEvents el name
 
 -- | Make a new DOM element with a given tag name.
@@ -268,6 +269,12 @@ instance MonadIO UI where
 
 instance MonadFix UI where
     mfix f = UI $ mfix (unUI . f)
+
+instance MonadThrow UI where
+    throwM = UI . throwM
+
+instance MonadCatch UI where
+    catch m f = UI $ catch (unUI m) (unUI . f)
 
 -- | Execute an 'UI' action in a particular browser window.
 -- Also runs all scheduled 'IO' actions.